+++ Swollball/GameHub.Player.cs
@@ -1,0 +1,8 @@
+﻿using Microsoft.AspNetCore.SignalR;
+
+namespace Swollball
+{
+    public partial class GameHub : Hub
+    {
+    }
+}
