+++ wwwroot/Swollball/Swollball_Host.html
@@ -1,7 +1,6 @@
 ﻿<!DOCTYPE html>
 <html>
 <head>
-
     <script src="/lib/signalr/dist/browser/signalr.js"></script>
     <script src="/phaser/phaser.js"></script>
     <script src="/js/app.js"></script>
@@ -30,7 +29,8 @@
 
 </body>
 </html>
-<script src="/gamelobby.js"></script>
+<script src="common.js"></script>
+<script src="gamelobby.js"></script>
 
 <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.2.1/css/bootstrap.min.css" integrity="sha384-GJzZqFGwb1QTTN6wy59ffF1BuGJpLSa9DkKMp0DgiMDm4iYMj70gZWKYbI706tWS" crossorigin="anonymous">
 <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.2.1/js/bootstrap.min.js" integrity="sha384-B0UglyR+jN6CkvvICOB2joaf5I4l3gm9GU6Hc1og6Ls7i6U/mkkaduKaBhlAXv9k" crossorigin="anonymous"></script>
