+++ wwwroot/index.html
@@ -5,6 +5,6 @@
     <title></title>
 </head>
 <body>
-    <a href="Swollball_Host.html">SWOLLBALL HOST GAME</a>
+    <a href="Swollball\Swollball_Host.html">SWOLLBALL HOST GAME</a>
 </body>
 </html>
